made change color, add automation to icon in storybook
