+++ src/shared/ui/icon/Icon.tsx
@@ -9,14 +9,19 @@ interface IconProps {
 	name: IconName
 	className?: string
 	style?: CSSProperties
+	color?: string
 }
-export const Icon = ({ name, className, style }: IconProps) => {
+export const Icon = ({ name, className, color }: IconProps) => {
 	const Svg = getIcon(name)
 
 	return (
 		<Svg
-			style={style}
-			className={classNames(cl.icon, className ? className : '')}
+			color={color}
+			className={classNames(
+				cl.root,
+				color ? cl.root_color : '',
+				className ?? ''
+			)}
 		/>
 	)
 }
