+++ src/shared/ui/icon/icon.module.scss
@@ -1,3 +1,7 @@
-.icon {
+.root {
 	display: inline-block;
+
+	&_color [fill] {
+		fill: currentcolor;
+	}
 }
